reorganized api routes, student model, and api controllers, added time file, changed time for token
