+++ util/time.js
@@ -1,0 +1,8 @@
+export const timeStamp = () => {
+  return Number(Date.now());
+};
+
+export const elapsedTime = (start, end) => {
+  const totalSeconds = (end - start) / 1000;
+  return totalSeconds;
+};
